fix: Clarify error offset in file-validate (#218)
fix: Clarify error offset of file-validate

Current version of the code, when showing the error offset in a
file-validate check many times say a wrong offset inside the file. This
is because it is showing the error offset inside the buffer reading
from the checked file.

The real offset may be obtained from the expected data, but it would be
much easier to show it as a first result.