feat(chunkserver): Use aligned memory for writes
The memory for write operations is now aligned to 4 KiB, which allows
to later implement disks writes in DirectIO. It also benefits SMR
writes by removing the need of memcpy to align the buffer to the
underlying drive block size.

The previous approach was also resizing constantly the vector
PacketStruct::packet, which implies frequent allocations. The new
PacketStruct::alignedBuffer is now reused most of the time without
constant allocations.