feat(chunkserver): Stat chunks at scan optionally (#245)
Scanning disks with millions of chunks can take a considerable time.
At scan time, by default, the chunkserver performs a stat operation on
the metadata and data files for each chunk. This check is good, but
makes the scan process very slow.

This change adds the configuration option STAT_CHUNKS_AT_DISK_SCAN to
optionally disable this behavior. The scan time with this option
disabled is around 40 times faster.