build: Get Prometheus externally
Previously it required you to have the library installed on the system.
Using FetchContent in CMake, it is now acquired externally if the CMake
config doesnot exist on the system and the build option is correctly set
(which it is by default). This will statically compile the library into
the binary.