test(all): Add new basic helgrind test
The threading safety was not checked explicitely in our tests.
Threading issues can cause non-deterministic runtime errors that are
hard to reproduce and debug. We use multiple threads so far in the
chunkservers and clients.

This change introduces a basic test using helgrind (from valgrind). The
idea is to gradually add complexity to this test and create others
(more specific) when needed.

The initial suppressions file contains placeholders that must be
classified to determine if they will remain as suppressions or a fix is
really needed.

The helgrind's history-level was additionally set to ```full``` for
more accurate catches.