chore: Add migrations structure
This change targets providing a place to store everything related to
migrations, which is specially important if there are breaking changes
involved. The idea is to fill the summary of the breaking changes and
the scripts for downgrading a specific service with the upcoming
breaking changes. The structure is also helpful for the next versions
upgrade scheme.

Changes:
- added a folder in the top level for the migration tools.
- added a folder inside migrations for the upgrade from v4.X.X to
v5.0.0, and another one inside of it for downgrade tools.
- added Breaking-changes-summary.md.
- added downgrade-master.sh which is the only service that needs a
special downgrade behavior for now.

Signed-off-by: Dave <dave@leil.io>