fix(mount): Reduce data race conditions in mount
Changes:
- use gReadCacheMemoryMutex instead of gUsedReadCacheMemoryMutex in
readdata_cache.h. This change unifies the protection on
gUsedReadCacheMemory.
- simplify the mutex protecting ReadCache::Entry to a regular
std::mutex.
- add the sequence lock-unlock-destory to the own mutex of
ReadCache::Entry and ReadRecord before releasing it. Apply similar
approach for threc.
- add lock when releasing a ReadCache::Result, for each of the released
entries. This one is not needed but will reduce data race warnings.
- add a lock when clearing a expired but not ready to be deleted
ReadRecord. This protection is needed, but was missing.
- make the decrease of requestsNotDone just before leaving the scope in
the read_worker, this is needed to make expired ReadRecords to be
deleted safely.
- make the release of the ownership of the ReadaheadRequestEntry inside
an inodeLock protected block of code. This is not needed, but will
further reduce data races warnings.
- unify the naming of the locks in read algorithm.
- move assignments to use_flocks and use_posixlocks in
Sauna_Client::release to remove_file_info to reduce data race warnings.