chore: Change how PACKAGE_VERSION string is build (#232)
* chore: Change how PACKAGE_VERSION string is build

Git commits are not guaranteed to be available or reliable on build time
(for example, some packaging systems prefer without a VCS or have their
own image of the repository with their own commits). Build time may be
useful, but it should not be part of the PACKAGE_VERSION, instead it
should be generated seperately and used seperately when -v is called.
Also, the version should not depend on an outside variable, the
CMakeLists.txt file should be the final source for this.

This commit removes the above git commit and build time variables (build
time will be re-introduced in another commit) and simplifies the version
string, with one exception: A compile time option called VERSION_SUFFIX,
which can be used to identify the build (e.g, 'unofficial', 'dev', 'rc'
or nothing at all). By default this will be dev when building from
source, or 'unofficial' when using ./configure. Package maintainers may
use their own suffix to distinguish building processes or add additional
version information.

- Add a new common function called version
- Add build time to version output again
- Add optional compile time definitions for specifing Git commit and Git
  branch
- Add missing version to rest of binaries