feat: Reduce malloc virtual memory usage
The new option LIMIT_GLIBC_MALLOC_ARENAS allows to tune (from
configuration) the number of arenas used by malloc in environments with
glibc. Setting it to a small positive number like 4 or 8, reduce the
virtual memory used by the application, at the cost of possible
performance drop for multi-threading (due to contention).

The default is 0, which means disabled (lets glibc decide the values).
The options is meant to be used in environments with limited RAM.

Signed-off-by: guillex <guillex@leil.io>