feat: Reuse the buffers when storing metadata
The goal of this change is to speed up the storing of metadata by
saving the allocation of the buffers used to write the nodes and edges
metadata before writing those into the metadata file. Also some more
explicit naming of the length constants were used to make code more
readable.

Signed-off-by: Dave <dave@leil.io>