feat: Reduce unlink overhead
Currently unlink syscall was calling several times the function
FSNodeDirectory::getChildName. That function was taking time
proportional to the number of entries inside the directory, which can
make the unlink operation in big directories to be very expensive.

This commit makes those operations taking time proportional to the
number of parents of the node to get the name of. The idea is to store,
beside the inode of the parents of a node, also a pointer to the handle
of the name of the node inside the parent entries structure.

As a side change, the case insensitive directories were optimized to be
O(logn) in the find queries instead of previous O(n).

Signed-off-by: Dave <dave@leil.io>