feat: Improve metadata speed for big directories (#272)
This commit changes the old flat_map that was advised to use for an
ordered_statistics structure that ensures good performance in every
case. The custom flat_map structure previously used had O(n) insertion
and elimination costs, the new one has O(logn) cost for both of those
operations. The goal of this change is to speedup the metadata
operations for directories with lots of files.

As a side effect, the compiler directive
DISABLE_JUDY_FOR_ENTRIESCONTAINER was removed, as it was not being used
at all.

Signed-off-by: Dave <dave@leil.io>