feat(mds,common): Start decoupling metadata backend (#296)
The metadata persistency layer (or backend) is hard-coded to use only
one way, mostly based on the metadata.sfs file. Implementing other
backends (like a database) is extremely complex, due to the internal
dependencies on this fixed format.

This commit starts a series of commits to decouple the metadata
persistency layer from the unique current implementation. The goal is
to make easy the implementation of new approaches.

The main changes of this specific commit are:
- Create the IMetadataBackend interface to provide the extension point.
- Create the MetadataBackendFile as concrete implementation, containing
  the current behavior.
- Remove the files filesystem_store.h and filesystem_store.cc from
  master (mostly moved to the new files).
- Remove the metadata.h and metadata.cc files from common, they were
  doing a job very specific to metadata servers.
- Update all modules to use the new gMetadataBackend instance (this
  will hold the concrete backend when available).

This change involves modifications in master, shadows, metaloggers
and metarestore.

Signed-off-by: guillex <guillex@leil.io>