fix: Fix metarestore polluting stderr
Because of the changes to static linking of spdlog from vcpkg, it is
automatically adding a default logger, that logs to stdout. This is not
affecting main, because we removed all logs at setup. However, this was
not the case for metarestore, which means stdout was polluted with logs,
causing some tests to fail to properly parse files.

This change fixes that by dropping all logs on metarestore, before
adding stderr and syslog logs.

Signed-off-by: Urmas Rist <urmas@leil.io>