build: Add vcpkg support (#302)
* feat(build): Add presets for cmake with vcpkg

- Add presets for cmake with VCPKG.
- Add utility scripts for vcpkg setup and packages install/update.
- Set the packages to gather from vcpkg registry.
- Lock the dependency versions to be used for avoiding the
  consequences of undesired upgrades at unexpected moment.
- Add VCPKG usage to the documentation.
- Remove Hunter package manager
- Update gtests dependencies

Co-authored-by: guillex <guillex@leil.io>

Signed-off-by: Baldor <baldor@leil.io>

* fix: Update github action for vcpkg

Signed-off-by: Baldor <baldor@leil.io>

* fix: Fix compilation in Windows

Signed-off-by: Dave <dave@leil.io>

* fix(tests): Link to header version of spdlog and fmt

Also, specify the linking scope explicitly for other libs.

Signed-off-by: Baldor <baldor@leil.io>

* build: Fix wrong path for examples

Due to where project("saunafs") is defined, it's adding the project name
to the folder with doc. This wasn't an issue in the past because project
name was defined later after the doc folder variable was declared.

This hardcodes the doc folder to be always
$CMAKE_INSTALL_FULL_DATAROOTDIR/doc/.

Signed-off-by: Urmas Rist <urmas@leil.io>

* ci: Refactor CI
* Add Dockerfile from sfstests
* Seperate package installation script (for caching in Dockerfile)
* Use a fixed version of llvm.sh (don't wget it)
* Use vcpkg in Docker
* Disable caching CMake configuration
* Add disabled Ubuntu 22.04 tests

Signed-off-by: Urmas Rist <urmas@leil.io>

* ci: Fix run-unit-and-ganesha-tests with vcpkg and codespell

Signed-off-by: Urmas Rist <urmas@leil.io>

* ci: Add deployment pipeline

This will use our internal deployment pipeline to build the packages and
push to developer repository, but ONLY on the dev branch.

Signed-off-by: Urmas Rist <urmas@leil.io>

* chore: Fix spelling

Signed-off-by: Urmas Rist <urmas@leil.io>

* revert(build): Bring back link scope to defaults

The 0ee0ef67 commit introduced changes that are out of the scope of
this PR.

Signed-off-by: Baldor <baldor@leil.io>

* fix(build): Apply PR suggestions

Signed-off-by: Baldor <baldor@leil.io>

* fix(build): Add missing header

The build broke and could be run with the header-only version of spdlog,
but that was a fleeting patch. This commit fixes the cause of the
failure and reverts the patch.

Signed-off-by: Baldor <baldor@leil.io>

* fix: Fix metarestore polluting stderr

Because of the changes to static linking of spdlog from vcpkg, it is
automatically adding a default logger, that logs to stdout. This is not
affecting main, because we removed all logs at setup. However, this was
not the case for metarestore, which means stdout was polluted with logs,
causing some tests to fail to properly parse files.

This change fixes that by dropping all logs on metarestore, before
adding stderr and syslog logs.

Signed-off-by: Urmas Rist <urmas@leil.io>

---------

Signed-off-by: Baldor <baldor@leil.io>
Signed-off-by: Dave <dave@leil.io>
Signed-off-by: Urmas Rist <urmas@leil.io>
Co-authored-by: guillex <guillex@leil.io>
Co-authored-by: Baldor <baldor@leil.io>
Co-authored-by: Dave <dave@leil.io>